there's a new ValidatorUtil class. Modified all service layer's classes. In model added validation methods where is necessary

+++ src/main/java/am/aca/wftartproject/service/impl/ArtistServiceImpl.java
@@ -12,12 +12,14 @@ import org.apache.log4j.Logger;
 import java.sql.Connection;
 import java.sql.SQLException;
 
+import static am.aca.wftartproject.service.impl.validator.ValidatorUtil.isEmptyString;
+
 /**
  * Created by surik on 6/3/17
  */
 public class ArtistServiceImpl implements ArtistService {
 
-    private static final Logger LOGGER = Logger.getLogger(ArtistService.class);
+    private static final Logger LOGGER = Logger.getLogger(ArtistServiceImpl.class);
 
     private Connection conn = new DBConnection().getDBConnection(DBConnection.DBType.REAL);
     private ArtistDao artistDao = new ArtistDaoImpl(conn);
@@ -28,36 +30,42 @@ public class ArtistServiceImpl implements ArtistService {
 
     /**
      * @param artist
-     *
      * @see ArtistService#addArtist(Artist)
      */
     @Override
     public void addArtist(Artist artist) {
 
-        if (artist.getSpecialization() != null){
-
-            try{
-                artistDao.addArtist(artist);
-            }catch (DAOException e){
-                String error = "Failed to add Artist: %s";
-                LOGGER.error(String.format(error, e.getMessage()));
-                throw new ServiceException(String.format(error, e.getMessage()));
-            }
+        if (artist == null || !artist.isValidArtist()) {
+            LOGGER.error(String.format("Artist is invalid: %s", artist));
+            throw new ServiceException("Invalid artist");
         }
+
+        try {
+            artistDao.addArtist(artist);
+        } catch (DAOException e) {
+            String error = "Failed to add Artist: %s";
+            LOGGER.error(String.format(error, e.getMessage()));
+            throw new ServiceException(String.format(error, e.getMessage()));
+        }
+
 
     }
 
     /**
      * @param id
      * @return
-     *
      * @see ArtistService#findArtist(Long)
      */
     @Override
     public Artist findArtist(Long id) {
-        try{
+        if (id == null || id < 0) {
+            LOGGER.error(String.format("Id is invalid: %s", id));
+            throw new ServiceException("Invalid id");
+        }
+
+        try {
             return artistDao.findArtist(id);
-        }catch (DAOException e){
+        } catch (DAOException e) {
             String error = "Failed to find Artist: %s";
             LOGGER.error(String.format(error, e.getMessage()));
             throw new ServiceException(String.format(error, e.getMessage()));
@@ -67,14 +75,18 @@ public class ArtistServiceImpl implements ArtistService {
     /**
      * @param email
      * @return
-     *
      * @see ArtistService#findArtist(String)
      */
     @Override
     public Artist findArtist(String email) {
-        try{
+        if (isEmptyString(email)) {
+            LOGGER.error(String.format("Email is invalid: %s", email));
+            throw new ServiceException("Invalid email");
+        }
+
+        try {
             return artistDao.findArtist(email);
-        }catch (DAOException e){
+        } catch (DAOException e) {
             String error = "Failed to find Artist: %s";
             LOGGER.error(String.format(error, e.getMessage()));
             throw new ServiceException(String.format(error, e.getMessage()));
@@ -84,14 +96,22 @@ public class ArtistServiceImpl implements ArtistService {
     /**
      * @param id
      * @param artist
-     *
      * @see ArtistService#updateArtist(Long, Artist)
      */
     @Override
     public void updateArtist(Long id, Artist artist) {
+        if (id == null || id < 0){
+            LOGGER.error(String.format("Id is invalid: %s", id));
+            throw new ServiceException("Invalid id");
+        }
+        if (artist == null || !artist.isValidArtist()){
+            LOGGER.error(String.format("Artist is invalid: %s", id));
+            throw new ServiceException("Invalid artist");
+        }
+
         try {
             artistDao.updateArtist(id, artist);
-        }catch (DAOException e){
+        } catch (DAOException e) {
             String error = "Failed to update Artist: %s";
             LOGGER.error(String.format(error, e.getMessage()));
             throw new ServiceException(String.format(error, e.getMessage()));
@@ -100,15 +120,19 @@ public class ArtistServiceImpl implements ArtistService {
 
     /**
      * @param id
-     *
      * @see ArtistService#deleteArtist(Long)
      */
     @Override
     public void deleteArtist(Long id) {
+        if (id == null || id < 0){
+            LOGGER.error(String.format("Id is invalid: %s", id));
+            throw new ServiceException("Invalid id");
+        }
+
         try {
             artistDao.deleteArtist(id);
 //        userDao.deleteUser(id);
-        }catch (DAOException e){
+        } catch (DAOException e) {
             String error = "Failed to delete Artist: %s";
             LOGGER.error(String.format(error, e.getMessage()));
             throw new ServiceException(String.format(error, e.getMessage()));
