+++ src/main/java/am/aca/wftartproject/service/impl/PurchaseHistoryServiceImpl.java
@@ -1,5 +1,7 @@
 package am.aca.wftartproject.service.impl;
 
+import am.aca.wftartproject.exception.DAOException;
+import am.aca.wftartproject.exception.ServiceException;
 import am.aca.wftartproject.model.PurchaseHistory;
 import am.aca.wftartproject.service.PurchaseHistoryService;
 import am.aca.wftartproject.util.DBConnection;
@@ -7,13 +9,14 @@ import am.aca.wftartproject.dao.PurchaseHistoryDao;
 import am.aca.wftartproject.dao.impl.PurchaseHistoryDaoImpl;
 import java.sql.SQLException;
 import java.util.List;
+import org.apache.log4j.Logger;
 
 /**
  * Created by surik on 6/1/17
  */
 public class PurchaseHistoryServiceImpl implements PurchaseHistoryService {
 
-
+    private static final Logger LOGGER = Logger.getLogger(PurchaseHistoryServiceImpl.class);
     private PurchaseHistoryDao purchaseHistoryDao = null;
 
 
@@ -32,8 +35,17 @@ public class PurchaseHistoryServiceImpl implements PurchaseHistoryService {
     @Override
     public void addPurchase(PurchaseHistory purchaseHistory) {
 
-        purchaseHistoryDao.addPurchase(purchaseHistory);
-
+        if (purchaseHistory == null || !purchaseHistory.isValidPurchaseHistroy()) {
+            LOGGER.error(String.format("purchaseHistory is invalid: %s", purchaseHistory));
+            throw new ServiceException("Invalid purchaseHistory");
+        }
+        try {
+            purchaseHistoryDao.addPurchase(purchaseHistory);
+        } catch (DAOException e) {
+            String error = "Failed to add purchase history: %s";
+            LOGGER.error(String.format(error, e.getMessage()));
+            throw new ServiceException(String.format(error, e.getMessage()));
+        }
     }
 
     /**
@@ -46,7 +58,21 @@ public class PurchaseHistoryServiceImpl implements PurchaseHistoryService {
     @Override
     public PurchaseHistory getPurchase(Long userId, Long itemId) {
 
-        return purchaseHistoryDao.getPurchase(userId, itemId);
+        if (userId == null || userId < 0){
+            LOGGER.error(String.format("userId is invalid: %s", userId));
+            throw new ServiceException("Invalid userId");
+        }
+        if (itemId == null || itemId < 0) {
+            LOGGER.error(String.format("itemId is invalid: %s", itemId));
+            throw new ServiceException("Invalid itemId");
+        }
+        try {
+            return purchaseHistoryDao.getPurchase(userId, itemId);
+        }catch (DAOException e) {
+            String error = "Failed to get purchase history: %s";
+            LOGGER.error(String.format(error, e.getMessage()));
+            throw new ServiceException(String.format(error, e.getMessage()));
+        }
     }
 
     /**
@@ -58,7 +84,17 @@ public class PurchaseHistoryServiceImpl implements PurchaseHistoryService {
     @Override
     public List<PurchaseHistory> getPurchase(Long userId) {
 
-        return purchaseHistoryDao.getPurchase(userId);
+        if (userId == null || userId < 0){
+            LOGGER.error(String.format("userId is invalid: %s", userId));
+            throw new ServiceException("Invalid userId");
+        }
+        try {
+            return purchaseHistoryDao.getPurchase(userId);
+        }catch (DAOException e) {
+            String error = "Failed to get purchase history: %s";
+            LOGGER.error(String.format(error, e.getMessage()));
+            throw new ServiceException(String.format(error, e.getMessage()));
+        }
     }
 
     /**
@@ -70,7 +106,21 @@ public class PurchaseHistoryServiceImpl implements PurchaseHistoryService {
     @Override
     public void deletePurchase(Long userId, Long itemId) {
 
-        purchaseHistoryDao.deletePurchase(userId, itemId);
+        if (userId == null || userId < 0){
+            LOGGER.error(String.format("userId is invalid: %s", userId));
+            throw new ServiceException("Invalid userId");
+        }
+        if (itemId == null || itemId < 0) {
+            LOGGER.error(String.format("itemId is invalid: %s", itemId));
+            throw new ServiceException("Invalid itemId");
+        }
+        try {
+            purchaseHistoryDao.deletePurchase(userId, itemId);
+        }catch (DAOException e) {
+            String error = "Failed to delete purchase history: %s";
+            LOGGER.error(String.format(error, e.getMessage()));
+            throw new ServiceException(String.format(error, e.getMessage()));
+        }
     }
 
 }
